requirements.txt
Flask==2.2.2
numpy==1.26.4
opencv_contrib_python==4.7.0.68
opencv_python==4.6.0.66
Pillow==10.3.0
setuptools==58.1.0
gunicorn==22.0.0
Werkzeug==2.2.2

